fix shitty bytecode extraction

+++ byond-extools/src/core/proc_management.h
@@ -41,8 +41,8 @@ namespace Core
 		ProcSetupEntry* setup_entry_bytecode = nullptr;
 		ProcSetupEntry* setup_entry_varcount = nullptr;
 
-		std::uint16_t bytecode_idx = 0;
-		std::uint16_t varcount_idx = 0;
+		std::uint32_t bytecode_idx = 0;
+		std::uint32_t varcount_idx = 0;
 
 		std::uint32_t* original_bytecode_ptr = nullptr;
 		std::vector<std::uint32_t> bytecode;
